refactor: move canvas surface under editor/tabs/architecture/
Relocate all canvas UI components into the editor/tabs/architecture
directory with kebab-case naming. This completes the stacks editor
restructuring to consolidate all editor surfaces under tabs/
