+++ frontend/src/pages/stacks/components/editor/tabs/architecture/add-resource-popover.tsx
