+++ frontend/src/pages/stacks/components/editor/tabs/architecture/canvas-context-menu.tsx
