+++ frontend/src/pages/stacks/components/editor/tabs/architecture/canvas-controls.tsx
