+++ frontend/src/pages/stacks/components/editor/tabs/architecture/drawer-stack.tsx
