+++ frontend/src/pages/stacks/components/editor/tabs/architecture/edges/connection-edge.tsx
