+++ frontend/src/pages/stacks/components/editor/tabs/architecture/fit-options.ts
