+++ frontend/src/pages/stacks/components/editor/tabs/architecture/mount-path-dialog.tsx
