+++ frontend/src/pages/stacks/components/editor/tabs/architecture/nodes/attachment-node.tsx
