+++ frontend/src/pages/stacks/components/editor/tabs/architecture/nodes/node-glyph.tsx
