+++ frontend/src/pages/stacks/components/editor/tabs/architecture/nodes/resource-node.tsx
