+++ frontend/src/pages/stacks/components/editor/tabs/architecture/resource-drawer.tsx
