+++ frontend/src/pages/stacks/components/editor/tabs/architecture/volume-drawer.tsx
